refactor: Move getErrorMessage in helpers
Also remove old import from a test file.
The test file itself seems a bit outdated,
it would be good to review it at some point
