--- src/drive/web/modules/drive/helpers.js
@@ -1,9 +1,0 @@
-import { makeStyles } from 'cozy-ui/transpiled/react/styles'
-
-export const useFabStyles = makeStyles(() => ({
-  root: {
-    position: 'fixed',
-    right: ({ right }) => (right ? right : '1rem'),
-    bottom: ({ bottom }) => (bottom ? bottom : '1rem')
-  }
-}))
